splitting up and genericizing some code, localizing default-dependent stuff in one file
